rename contracts

+++ contracts/Mocks/IUSDT.sol
@@ -1,11 +1,15 @@
 pragma solidity ^0.5.8;
 
 contract ERC20Basic {
-    uint public _totalSupply;
-    function totalSupply() public view returns (uint);
-    function balanceOf(address who) public view returns (uint);
-    function transfer(address to, uint value) public;
-    event Transfer(address indexed from, address indexed to, uint value);
+  uint256 public _totalSupply;
+
+  function totalSupply() public view returns (uint256);
+
+  function balanceOf(address who) public view returns (uint256);
+
+  function transfer(address to, uint256 value) public;
+
+  event Transfer(address indexed from, address indexed to, uint256 value);
 }
 
 /**
@@ -13,8 +17,18 @@ contract ERC20Basic {
  * @dev see https://github.com/ethereum/EIPs/issues/20
  */
 contract IUSDT is ERC20Basic {
-    function allowance(address owner, address spender) public view returns (uint);
-    function transferFrom(address from, address to, uint value) public;
-    function approve(address spender, uint value) public;
-    event Approval(address indexed owner, address indexed spender, uint value);
+  function allowance(address owner, address spender)
+    public
+    view
+    returns (uint256);
+
+  function transferFrom(
+    address from,
+    address to,
+    uint256 value
+  ) public;
+
+  function approve(address spender, uint256 value) public;
+
+  event Approval(address indexed owner, address indexed spender, uint256 value);
 }
